pyproject.toml
[project]
name = "dbms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-connector-python>=9.4.0",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.50.0",
]
